requirements.txt
pyqt6
pyaudio
numpy
openai-whisper
pyinstaller
requests
